refactor: move inputs and outputs to separate file
