+++ internal/api/hook_inputs.go
@@ -1,0 +1,6 @@
+package api
+
+// TODO: Get staticcheck to ignore this
+// type CustomSMSSenderRequest struct {
+// 	phone string
+// }
